moved form and input components to common folder
